+++ src/main/kotlin/racingcar/model/car/Car.kt
@@ -10,15 +10,7 @@ class Car(name: String, private var _position: Int = 0) {
         }
     }
 
-    fun toDto(): CarDto = CarDto(carName.name)
-
     companion object {
         private const val MOVEMENT_PROBABILITY = 4
     }
-}
-
-class CarDto(_carName: String, val position: Int = 0) {
-    val carName: CarName = CarName(_carName.trim())
-
-    fun toModel(): Car = Car(carName.name, position)
 }
