+++ src/test/kotlin/racingcar/service/RacingServiceTest.kt
@@ -6,7 +6,7 @@ import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments
 import org.junit.jupiter.params.provider.MethodSource
 import org.mockito.Mockito
-import racingcar.model.car.CarDto
+import racingcar.dto.car.CarDto
 import racingcar.model.car.CarsDto
 import racingcar.model.round.RoundDto
 import racingcar.utils.random.MovementProbabilityGenerator
